### 2017-01-09 #156 @whedon
[PRE REVIEW]: biotmle: Targeted Learning for Biomarker Discovery
**Submitting author:** @nhejazi (<a href="http://orcid.org/0000-0002-7127-2789">Nima Hejazi</a>)
**Repository:** <a href="https://github.com/nhejazi/biotmle" target ="_blank">https://github.com/nhejazi/biotmle</a>
**Version:** v0.1.0
