### 2017-01-11 #156 @arfon
👋 @mgymrek @biorelated - would either of you be willing to edit this submission for JOSS?
