### 2017-01-11 c36301c
update pkgdown site, minor travis/appveyor revisions
