+++ appveyor.yml
@@ -15,7 +15,7 @@ cache:
 # Adapt as necessary starting from here
 environment:
   global:
-    WARNINGS_ARE_ERRORS: 0
+    WARNINGS_ARE_ERRORS: 1
 
   matrix:
   - TEST_DIR: pkg
@@ -29,6 +29,7 @@ before_build:
 build_script:
   - travis-tool.sh install_deps
   - travis-tool.sh install_bioc limma
+  - travis-tool.sh install_bioc_deps
 
 test_script:
   - travis-tool.sh run_tests
