+++ README.md
@@ -2,7 +2,7 @@
 
 [![Travis-CI Build Status](https://travis-ci.org/nhejazi/biotmle.svg?branch=master)](https://travis-ci.org/nhejazi/biotmle)
 [![AppVeyor Build  Status](https://ci.appveyor.com/api/projects/status/github/nhejazi/biotmle?branch=master&svg=true)](https://ci.appveyor.com/project/nhejazi/biotmle)
-[![Coverage Status](https://coveralls.io/repos/github/nhejazi/biotmle/badge.svg?branch=master)](https://coveralls.io/github/nhejazi/biotmle?branch=master)
+[![Coverage Status](https://coveralls.io/repos/github/nhejazi/biotmle/badge.svg)](https://coveralls.io/github/nhejazi/biotmle)
 [![CRAN](http://www.r-pkg.org/badges/version/biotmle)](http://www.r-pkg.org/pkg/biotmle)
 [![Project Status: WIP - Initial development is in progress, but there has not yet been a stable, usable release suitable for the public.](http://www.repostatus.org/badges/latest/wip.svg)](http://www.repostatus.org/#wip)
 [![MIT license](http://img.shields.io/badge/license-MIT-brightgreen.svg)](http://opensource.org/licenses/MIT)
