### 2017-01-14 6d3838e
add initial Bioc compliance
