+++ pkg/NEWS.md
@@ -1,2 +1,0 @@
-February 1, 2017
-* The first working releast of package `tmlelimma` is available on CRAN.
