+++ pkg/DESCRIPTION
@@ -1,16 +1,16 @@
 Package: biotmle
 Title: Targeted Learning for Biomarker Discovery with the Moderated T-Statistic
-Version: 0.1.0
+Version: 0.99.0
 Authors@R: as.person(c(
-    "Nima S. Hejazi <nhejazi@berkeley.edu> [aut, cre]",
-    "Alan E. Hubbard <hubbard@berkeley.edu> [aut]",
+    "Nima S Hejazi <nhejazi@berkeley.edu> [aut, cre]",
+    "Alan E Hubbard <hubbard@berkeley.edu> [aut]",
     "Weixin Cai <wcai@berkeley.edu> [ctb]"))
-Description: This package facilitates the discovery of biomarkers from Omics
-    data based on the associations of potential biomarkers with exposure and
-    outcome variables by implementing an estimation procedure that combines a
-    generalization of the moderated t-statistic with asymptotically linear
-    statistical parameters estimated via targeted minimum loss-based estimation
-    (TMLE).
+Description: This package facilitates the discovery of biomarkers from
+    biological sequencing data (e.g., microarrays, RNA-seq) based on the
+    associations of potential biomarkers with exposure and outcome variables by
+    implementing an estimation procedure that combines a generalization of the
+    moderated t-statistic with asymptotically linear statistical parameters
+    estimated via targeted minimum loss-based estimation (TMLE).
 Depends:
     R (>= 3.3.1)
 License: MIT + file LICENSE
