### 2017-01-17 496b13a
numerous updates for BioC
