+++ pkg/R/biotmle.R
@@ -8,9 +8,15 @@
 #' @param tmleOut \code{data.frame} containing the output generated by use of
 #'        \code{biomarkertmle}
 #'
-#' @return S3 object of class "biotmle" for TMLE-based biomarker analysis.
+#' @return S3 object of class \code{biotmle} for TMLE-based biomarker analysis.
 #'
 #' @export biotmle
+#'
+#' @examples
+#' biotmle <- biotmle(call = NULL,
+#'                    topTable = NULL,
+#'                    limmaOut = NULL,
+#'                    tmleOut = NULL)
 #'
 biotmle <- function(call = NULL,
                     topTable = NULL,
