+++ pkg/R/limmatmle.R
@@ -1,0 +1,66 @@
+#' The Moderated T-Statistic for Asymptotically Linear Parameters
+#'
+#' Performs variance shrinkage via empirical Bayes procedures on influence
+#' curve transforms of the Average Treatment Effect of gene expression
+#' changes on the outcome of interest
+#'
+#' @param biotmle \code{biotmle} object as generated by \code{biomarkertmle}
+#' @param designMat a design matrix providing the contrasts to be used in the
+#'        linear model fitting procedure of \code{limma::lmFit}
+#' @param ... additional arguments to be passed to functions from \code{limma}
+#'
+#' @importFrom limma lmFit eBayes topTable
+#'
+#' @return \code{biotmle} object containing output from \code{limma::lmFit} and
+#'         \code{limma::topTable}
+#'
+#' @export limmatmle
+#'
+#' @examples
+#' library(dplyr)
+#' data(illuminaData)
+#' data(biomarkertmleOut)
+#' "%ni%" = Negate("%in%")
+#'
+#' W <- illuminaData %>%
+#'  dplyr::select(which(colnames(.) %in% c("age", "sex", "smoking"))) %>%
+#'  dplyr::mutate(
+#'    age = as.numeric((age > quantile(age, 0.25))),
+#'    sex = I(sex),
+#'    smoking = I(smoking)
+#'  )
+#'
+#' A <- illuminaData %>%
+#'  dplyr::select(which(colnames(.) %in% c("benzene")))
+#' A <- A[, 1]
+#'
+#' Y <- illuminaData %>%
+#'  dplyr::select(which(colnames(.) %ni% c("age", "sex", "smoking", "benzene",
+#'                                         "id")))
+#' geneIDs <- colnames(Y)
+#'
+#' design <- as.data.frame(cbind(rep(1, nrow(Y)),
+#'                               as.numeric(A == max(unique(A)))))
+#' colnames(design) <- c("intercept", "Tx")
+#'
+#' limmaTMLEout <- limmatmle(biotmle = biomarkerTMLEout, IDs = NULL,
+#'                           designMat = design)
+#'
+limmatmle <- function(biotmle,
+                      designMat,
+                      ...) {
+
+  biomarkerTMLEout <- biotmle$tmleOut
+
+  fit <- limma::lmFit(as.data.frame(biomarkerTMLEout), designMat)
+  fit <- limma::eBayes(fit)
+
+  tt <- limma::topTable(fit, coef = 2, adjust.method = "BH", sort.by = "none",
+                        number = Inf)
+  tt$IDs <- rownames(biomarkerTMLEout)
+
+  biotmle$topTable <- tt
+  biotmle$limmaOut <- fit
+
+  return(biotmle)
+}

[…]
