+++ NAMESPACE
@@ -34,5 +34,3 @@ importFrom(magrittr,"%>%")
 importFrom(parallel,detectCores)
 importFrom(tmle,tmle)
 importFrom(wesanderson,wes_palette)
-import(SuperLearner)
-import(Matrix)
