+++ .travis.yml
@@ -3,6 +3,11 @@ branches:
   - master
   - develop
 
+env:
+  global:
+    - R_BUILD_ARGS="--no-build-vignettes --no-manual"
+    - R_CHECK_ARGS="--no-build-vignettes --no-manual"
+
 language: r
 
 sudo: required
@@ -12,10 +17,13 @@ cache: packages
 warnings_are_errors: true
 
 r:
-  - bioc-release
-  - bioc-devel
   - release
   - devel
+
+before_install:
+  - curl -OL http://raw.github.com/craigcitro/r-travis/master/scripts/travis-tool.sh
+  - chmod 755 ./travis-tool.sh
+  - ./travis-tool.sh bootstrap
 
 install:
   - ./travis-tool.sh install_bioc limma
