### 2017-01-18 a3d0cfd
more BioC updates
