### 2017-01-18 3ac419c
rm lfs data, update data-gen script, misc
