+++ data-raw/illuminaData.R
@@ -1,6 +1,7 @@
+library(here)
 "%ni%" = Negate("%in%")
 
-data <- illumina2007 %>%
+illuminaData <- illumina2007 %>%
   dplyr::select(which(colnames(.) %ni% c("box", "riboaveugml", "ng", "exclude",
                                          "hyb", "totalrnaug", "chip", "Chip.Id",
                                          "Chip.Section", "label.c", "benzene",
@@ -12,3 +13,6 @@ data <- illumina2007 %>%
     smoking = I(current_smoking)
   ) %>%
   dplyr::select(which(colnames(.) %ni% c("newbenz", "current_smoking")))
+
+save(illuminaData, file = paste0(normalizePath(here("..", "data")),
+                                 "illuminaData.rda"))
