### 2017-01-18 2c62481
fix data-gen script, add illuminaData
