+++ data-raw/illuminaData.R
@@ -15,4 +15,4 @@ illuminaData <- illumina2007 %>%
   dplyr::select(which(colnames(.) %ni% c("newbenz", "current_smoking")))
 
 save(illuminaData, file = paste0(normalizePath(here("..", "data")),
-                                 "illuminaData.rda"))
+                                 "/illuminaData.rda"))
