### 2017-01-19 daa2a5f
minor plot/vignette corrections, add biomarkertmle data
