+++ vignettes/exposureBiomarkers.Rmd
@@ -145,17 +145,17 @@ for controlling the False Discovery Rate) as generated by `limma::topTable()`:
 plot(x = limmaTMLEout, type = "pvals_adj")
 ```
 
-Setting the argument `type="pvals_raw"` will instead produce a histogram of the
-raw p-values _(these are less informative and should, in general, not be used
-for inferential purposes, as the computation producing these p-values ignores
-the multiple testing nature of the biomarker discovery problem)_:
+Setting the argument `type = "pvals_raw"` will instead produce a histogram of
+the raw p-values _(these are less informative and should, in general, not be
+used for inferential purposes, as the computation producing these p-values
+ignores the multiple testing nature of the biomarker discovery problem)_:
 
 ```{r pval_hist_limma_rawp}
 plot(x = limmaTMLEout, type = "pvals_raw")
 ```
 
 Heatmaps are useful graphics for visualizing the relationship between measures
-on genomic objects and covariates of interest. The `aheatmap.biotmle()` function
+on genomic objects and covariates of interest. The `heatmap_biotmle()` function
 provides this graphic for `biotmle` objects, allowing for the relationship
 between the exposure variable and some number of "top" biomarkers (as
 determined by the call to `limmatmle()`) to be visualized. In general, the
@@ -174,7 +174,7 @@ The volcano plot is standard graphical tools for examining how changes in
 expression relate to the raw p-value. The utility of such plots lies in their
 providing a convenient way to identify and systematically ignore those genomic
 objects that have extremely low p-values due to extremely low variance between
-observations. The `volcanoPlot_biotmle()` function provides much of the same
+observations. The `volcano_biotmle()` function provides much of the same
 interpretation, except that the fold change values displayed in the x-axis refer
 to changes in the _contributions of each biomarker to the Average Treatment
 Effect_ (in standard practice, for microarray technology, these would be fold
