+++ man/biotmle.Rd
@@ -24,11 +24,4 @@ S3 object of class \code{biotmle} for TMLE-based biomarker analysis.
 \description{
 S3 constructor for class \code{biotmle}
 }
-\examples{
-biotmle <- biotmle(call = NULL,
-                   topTable = NULL,
-                   limmaOut = NULL,
-                   tmleOut = NULL)
-
-}
 
