+++ NAMESPACE
@@ -2,7 +2,6 @@
 
 S3method(plot,biotmle)
 export(biomarkertmle)
-export(biotmle)
 export(heatmap_biotmle)
 export(limmatmle)
 export(volcano_biotmle)
