### 2017-01-19 fe094a8
try new Travis format
