+++ .travis.yml
@@ -9,33 +9,22 @@ env:
     - R_CHECK_ARGS="--no-build-vignettes --no-manual"
 
 language: r
-
 sudo: required
-
 cache: packages
-
 warnings_are_errors: true
 
 r:
   - release
   - devel
 
-before_install:
-  - curl -OL http://raw.github.com/craigcitro/r-travis/master/scripts/travis-tool.sh
-  - chmod 755 ./travis-tool.sh
-  - ./travis-tool.sh bootstrap
-
-install:
-  - ./travis-tool.sh install_bioc limma
-  - ./travis-tool.sh install_bioc_deps
-
-## coverage
-r_packages:
-  - covr
+r_github_packages:
+  - jimhester/covr
 
 bioc_packages:
+  - limma
   - BiocCheck
 
+## coverage
 after_success:
   - Rscript -e 'covr::codecov()'
   - R CMD BiocCheck .
