### 2017-01-19 b05086d
resolve Bioc errors, re-trigger build
