+++ inst/doc/outcomeBiomarkers.R
@@ -1,4 +1,0 @@
-## ----setup_data----------------------------------------------------------
-library(dplyr)
-library(biotmle)
-
