+++ inst/doc/exposureBiomarkers.Rmd
@@ -1,186 +1,0 @@
----
-title: "Identifying Biomarkers from an Exposure Variable"
-author: "Nima S. Hejazi & Alan E. Hubbard"
-date: "`r Sys.Date()`"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Identifying Biomarkers from an Exposure Variable}
-  %\VignetteEngine{knitr::rmarkdown}
-  \usepackage[utf8]{inputenc}
----
-
-## Introduction
-
-The `biotmle` R package can be used to isolate biomarkers in two ways: based
-on the associations of genomic objects with (1) an exposure variable of interest
-and (2) an outcome variable of interest. In this vignette, we illustrate how to
-use `biotmle` to isolate and visualize genes associated with an __exposure__,
-using a data set containing microarray expression measures from an Illumina
-platform. In the analysis described below, Targeted Minimum Loss-Based
-Estimation (TMLE) is used to transform the microarray expression values based
-on the influence curve representation of the Average Treatment Effect (ATE).
-Following this transformation, the moderated t-statistic of Smyth is used to
-test for a binary groupwise difference (based on the exposure variable), using
-the tools provided by the R package `limma`.
-
----
-
-## Biomarker Identification
-
-First, we load the `biotmle` package and the (included) `illuminaData` data set:
-
-```{r setup_data}
-library(dplyr)
-library(biotmle)
-data(illuminaData)
-"%ni%" = Negate("%in%")
-subjIDs <- illuminaData$id
-```
-
-In order to perform Targeted Minimum Loss-Based Estimation, we need three
-separate data structures: (1) _W_, baseline covariates that could potentially
-confound the association of biomarkers with the exposure of interest; (2) _A_,
-the exposure of interest; and (3) _Y_, the biomarkers of interest. All values in
-_W_ and _A_ must be discretized. We demonstrate the process of creating these
-separate data structures with the `illuminaData` data set below:
-
-```{r clean_data}
-# W - age, sex, smoking
-W <- illuminaData %>%
-  dplyr::select(which(colnames(.) %in% c("age", "sex", "smoking"))) %>%
-  dplyr::mutate(
-    age = as.numeric((age > quantile(age, 0.25))),
-    sex = I(sex),
-    smoking = I(smoking)
-  )
-
-
-# A - benzene exposure (discretized)
-A <- illuminaData %>%
-  dplyr::select(which(colnames(.) %in% c("benzene")))
-A <- A[, 1]
-
-
-# Y - genes
-Y <- illuminaData %>%
-  dplyr::select(which(colnames(.) %ni% c("age", "sex", "smoking", "benzene",
-                                         "id")))
-geneIDs <- colnames(Y)
-```
-
-After creating _W_, _A_, and _Y_, the TMLE-based biomarker discovery process
-can be invoked using the `biomarkertmle` function. The procedure is quite
-resource-intensive because it evaluates the association of each individual
-potential biomarker (of which there are over 20,000 in the included data set)
-with exposure of interest _A_, while accounting for potential confounding based
-on the covariates included in _W_. We demonstrate the necessary syntax for
-calling `biomarkertmle` below:
-
-```{r biomarkerTMLE_eval, eval=FALSE}
-biomarkerTMLEout <- biomarkertmle(Y = Y,
-                                  W = W,
-                                  A = A,
-                                  type = "exposure",
-                                  parallel = TRUE,
-                                  family = "gaussian",
-                                  g_lib = c("SL.glmnet", "SL.randomForest",
-                                            "SL.polymars", "SL.mean"),
-                                  Q_lib = c("SL.glmnet", "SL.randomForest",
-                                            "SL.nnet", "SL.mean")
-                                 )
-```
-
-```{r load_biomarkerTMLE_result, echo=FALSE}
[…]
