+++ inst/doc/outcomeBiomarkers.Rmd
@@ -1,15 +1,0 @@
----
-title: "Identifying Biomarkers from an Outcome Variable"
-author: "Nima S. Hejazi & Alan E. Hubbard"
-date: "`r Sys.Date()`"
-output: rmarkdown::html_vignette
-vignette: >
-  %\VignetteIndexEntry{Identifying Biomarkers from an Outcome Variable}
-  %\VignetteEngine{knitr::rmarkdown}
-  \usepackage[utf8]{inputenc}
----
-
-```{r setup_data}
-library(dplyr)
-library(biotmle)
-```
