+++ NAMESPACE
@@ -33,3 +33,5 @@ importFrom(magrittr,"%>%")
 importFrom(parallel,detectCores)
 importFrom(tmle,tmle)
 importFrom(wesanderson,wes_palette)
+import(SuperLearner)
+import(Matrix)

[…]
