### 2017-01-19 1d96086
rm gitattributes, try Bioc trigger again
