+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: biotmle
 Title: Targeted Learning for Biomarker Discovery with the Moderated T-Statistic
-Version: 0.99.2
+Version: 0.99.3
 Authors@R: c(
     person("Nima", "Hejazi", email = "nhejazi@berkeley.edu",
            role = c("aut", "cre", "cph")),
