### 2017-02-06 nhejazi/biotmle#15 @nhejazi
TODO for Bioconductor
To-do list based on comments from the in-progress Bioconductor [review](https://github.com/Bioconductor/Contributions/issues/254#issuecomment-277282187):

* [x] switch `illuminaData` to use the SummarizedExperiment package and class
* [x] change `biomarkertmle` to accept input as objects of class SummarizedExperiment
* [x] re-write the `biotmle` class to simply exist as a subclass of SummarizedExperiment
* [x] explicitly check the "type" argument when first calling the `biomarkertmle` function
* [x] re-write heatmap function to make use of the SuperHeat package (see [this issue](https://github.com/nhejazi/biotmle/issues/14))
