### 2017-02-09 76d0bf4
correct install instructions
