+++ NAMESPACE
@@ -5,8 +5,6 @@ export(biomarkertmle)
 export(heatmap_biotmle)
 export(limmatmle)
 export(volcano_biotmle)
-importFrom(NMF,aheatmap)
-importFrom(NMF,nmf.options)
 importFrom(doParallel,registerDoParallel)
 importFrom(dplyr,arrange)
 importFrom(dplyr,filter)
@@ -31,7 +29,7 @@ importFrom(limma,lmFit)
 importFrom(limma,topTable)
 importFrom(magrittr,"%>%")
 importFrom(parallel,detectCores)
+importFrom(stats,quantile)
+importFrom(superheat,superheat)
 importFrom(tmle,tmle)
 importFrom(wesanderson,wes_palette)
-import(SuperLearner)
-import(Matrix)
