### 2017-03-13 #156 @nhejazi
Just wanted to add a few updates from the development end of the `biotmle` R package:

* There are a number of minor revisions (no changes to the core functionality) currently in process, for submission of the package to the [Bioconductor project](http://www.bioconductor.org/). See [this issue](https://github.com/nhejazi/biotmle/issues/15) for a list.

* All functionality of the package is intact -- perhaps see the R package vignette [here](http://nimahejazi.org/biotmle/articles/exposureBiomarkers.html).

* My understanding of the review process is that JOSS is meant to ensure that the software works as advertised, so I would think that deep knowledge of the statistical approach would be unnecessary (there is, of course, a separate methods manuscript being prepared for submission, but that should not be necessary to use the R package).
