### 2017-03-17 4613f12
pass in options to superheat properly
