### 2017-03-29 16fc629
main input to SumExp, rm data
