+++ DESCRIPTION
@@ -34,13 +34,16 @@ Imports:
     dplyr,
     superheat,
     SuperLearner,
-    Matrix
+    Matrix,
+    biotmleData
 Suggests:
     testthat,
     rmarkdown,
     knitr
+Remotes:
+    nhejazi/biotmleData
 VignetteBuilder: knitr
-RoxygenNote: 5.0.1
+RoxygenNote: 6.0.1
 biocViews:
     GeneExpression,
     DifferentialExpression,
