### 2017-03-29 2c43336
more minor BioC-rec fixes
