### 2017-04-03 119de84
major (back-incompat) changes for Bioc submission
