+++ R/modtest_ic.R
@@ -1,0 +1,54 @@
+#' Moderated Statistical Tests for Influence Curves
+#'
+#' Performs variance shrinkage via the empirical Bayes procedure of LIMMA on the
+#' observed data after a transformation moving the data to influence curve
+#' space, based on the average treatment effect parameter.
+#'
+#' @param biotmle \code{biotmle} object as generated by \code{biomarkertmle}
+#' @param design a design matrix providing the contrasts to be used in the
+#'        linear model fitting procedure of \code{limma::lmFit}
+#' @param ... additional arguments to be passed to functions from \code{limma}
+#'
+#' @importFrom limma lmFit eBayes topTable
+#'
+#' @return \code{biotmle} object containing output from \code{limma::lmFit} and
+#'         \code{limma::topTable}
+#'
+#' @export modtest_ic
+#'
+#' @examples
+#' library(dplyr)
+#' library(biotmleData)
+#' library(SummarizedExperiment)
+#' data(illuminaData)
+#' data(biomarkertmleOut)
+#' "%ni%" = Negate("%in%")
+#'
+#' colData(illuminaData) <- colData(illuminaData) %>%
+#'      data.frame %>%
+#'      dplyr::mutate(age = as.numeric(age > median(age))) %>%
+#'      DataFrame
+#'
+#' varInt_index <- which(names(colData(illuminaData)) %in% "benzene")
+#' designVar <- as.data.frame(colData(illuminaData))[, varInt_index]
+#' design <- as.numeric(designVar == max(designVar))
+#'
+#' limmaTMLEout <- modtest_ic(biotmle = biomarkerTMLEout, IDs = NULL,
+#'                           design = design)
+#'
+modtest_ic <- function(biotmle, design, ...) {
+
+  biomarkerTMLEout <- biotmle$tmleOut
+
+  fit <- limma::lmFit(as.data.frame(biomarkerTMLEout), design)
+  fit <- limma::eBayes(fit)
+
+  tt <- limma::topTable(fit, coef = 1, adjust.method = "BH", sort.by = "none",
+                        number = Inf)
+  tt$IDs <- rownames(biomarkerTMLEout)
+
+  biotmle$topTable <- tt
+  biotmle$limmaOut <- fit
+
+  return(biotmle)
+}
