+++ R/plots.R
@@ -19,32 +19,24 @@
 #'
 #' @examples
 #' library(dplyr)
+#' library(biotmleData)
+#' library(SummarizedExperiment)
 #' data(illuminaData)
 #' data(biomarkertmleOut)
 #' "%ni%" = Negate("%in%")
 #'
-#' W <- illuminaData %>%
-#'  dplyr::select(which(colnames(.) %in% c("age", "sex", "smoking"))) %>%
-#'  dplyr::mutate(
-#'    age = as.numeric((age > quantile(age, 0.25))),
-#'    sex = I(sex),
-#'    smoking = I(smoking)
-#'  )
+#' colData(illuminaData) <- colData(illuminaData) %>%
+#'      data.frame %>%
+#'      dplyr::mutate(age = as.numeric(age > median(age))) %>%
+#'      DataFrame
 #'
-#' A <- illuminaData %>%
-#'  dplyr::select(which(colnames(.) %in% c("benzene")))
-#' A <- A[, 1]
+#' varInt_index <- which(names(colData(illuminaData)) %in% "benzene")
+#' designVar <- as.data.frame(colData(illuminaData))[, varInt_index]
+#' design <- as.numeric(designVar == max(designVar))
 #'
-#' Y <- illuminaData %>%
-#'  dplyr::select(which(colnames(.) %ni% c("age", "sex", "smoking", "benzene",
-#'                                         "id")))
-#' geneIDs <- colnames(Y)
+#' limmaTMLEout <- modtest_ic(biotmle = biomarkerTMLEout, IDs = NULL,
+#'                            design = design)
 #'
-#' design <- as.data.frame(cbind(rep(1, nrow(Y)),
-#'                              as.numeric(A == max(unique(A)))))
-#' colnames(design) <- c("intercept", "Tx")
-#' limmaTMLEout <- limmatmle(biotmle = biomarkerTMLEout, IDs = NULL,
-#'                           designMat = design)
 #' plot(x = limmaTMLEout, type = "pvals_adj")
 #'
 plot.biotmle <- function(x, ..., type = "pvals_adj") {
@@ -99,38 +91,31 @@ plot.biotmle <- function(x, ..., type = "pvals_adj") {
 #'         the log-fold change in the causal target parameter against the raw
 #'         log p-value computed from the moderated t-test in \code{limmatmle}.
 #'
-#' @export volcano_biotmle
+#' @export volcano_ic
 #'
 #' @examples
 #' library(dplyr)
+#' library(biotmleData)
+#' library(SummarizedExperiment)
 #' data(illuminaData)
 #' data(biomarkertmleOut)
 #' "%ni%" = Negate("%in%")
 #'
-#' W <- illuminaData %>%
-#'  dplyr::select(which(colnames(.) %in% c("age", "sex", "smoking"))) %>%
-#'  dplyr::mutate(
-#'    age = as.numeric((age > quantile(age, 0.25))),
-#'    sex = I(sex),
-#'    smoking = I(smoking)
-#'  )
+#' colData(illuminaData) <- colData(illuminaData) %>%
+#'      data.frame %>%
+#'      dplyr::mutate(age = as.numeric(age > median(age))) %>%
+#'      DataFrame
 #'
-#' A <- illuminaData %>%
-#'  dplyr::select(which(colnames(.) %in% c("benzene")))
-#' A <- A[, 1]
+#' varInt_index <- which(names(colData(illuminaData)) %in% "benzene")
+#' designVar <- as.data.frame(colData(illuminaData))[, varInt_index]
+#' design <- as.numeric(designVar == max(designVar))
 #'
-#' Y <- illuminaData %>%
-#'  dplyr::select(which(colnames(.) %ni% c("age", "sex", "smoking", "benzene",
-#'                                         "id")))
[…]
