+++ R/zzz.R
@@ -1,5 +1,5 @@
 .onAttach <- function(...) {
-  packageStartupMessage("biotmle: Targeted Learning for Biomarker Discovery")
+  packageStartupMessage("biotmle: Moderated Statistics and Targeted Learning for Biomarker Discovery")
   packageStartupMessage("Version: ",
                         utils::packageDescription("biotmle")$Version)
 }

[…]
