### 2017-04-04 6186044
breaking changes: biotmle => S4
