+++ R/biotmle.R
@@ -1,28 +1,16 @@
-#' S3 constructor for class \code{biotmle}
+#' Constructor for class biotmle
 #'
-#' @param call merely passed in from \code{match.call(expand.dots = TRUE)}.
-#' @param topTable \code{data.frame} containing the output generated by use of
-#'        \code{limma::topTable}.
-#' @param limmaOut \code{MArrayList} containing the output generated by use of
-#'        \code{limma::lmFit}.
-#' @param tmleOut \code{data.frame} containing the output generated by use of
-#'        \code{biomarkertmle}
+#' @return class \code{biotmle} object, sub-classed from SummarizedExperiment.
 #'
-#' @return S3 object of class \code{biotmle} for TMLE-based biomarker analysis.
+#' @importClassesFrom SummarizedExperiment SummarizedExperiment
 #'
-biotmle <- function(call = NULL,
-                    topTable = NULL,
-                    limmaOut = NULL,
-                    tmleOut = NULL) {
-  # ============================================================================
-  # generic skeleton for S3 class
-  # ============================================================================
-  out <- structure(list(call,
-                        topTable,
-                        limmaOut,
-                        tmleOut),
-                   class = "biotmle")
-
-  names(out) <- c("call", "topTable", "limmaOut", "tmleOut")
-  return(out)
-}
+#' @export .biotmle
+#'
+.biotmle <- setClass(
+       Class = "bioTMLE",
+       slots = list(call = "call",
+                    tmleOut = "data.frame",
+                    modtestOut = "data.frame",
+                    topTable = "data.frame"),
+       contains = "SummarizedExperiment"
+)
