+++ R/biomarkertmle.R
@@ -29,10 +29,12 @@ utils::globalVariables("gene")
 #' @importFrom foreach foreach "%dopar%"
 #' @importFrom SummarizedExperiment assay colData
 #'
-#' @return S3 object of class \code{biotmle}, with the \code{tmleOut} and the
-#'         \code{call} slots containing TMLE-based estimates of the relationship
-#'         between a biomarker and exposure/outcome variable and the original
-#'         call to this function (for user reference), respectively.
+#' @return S4 object of class \code{biotmle}, generated by sub-classing
+#'         \code{SummarizedExperiment}, with additional slots containing
+#'         \code{tmleOut} and \code{call}, among others, containing TMLE-based
+#'         estimates of the relationship between a biomarker and exposure or
+#'         outcome variable and the original call to this function (for user
+#'         reference), respectively.
 #'
 #' @export biomarkertmle
 #'
@@ -81,12 +83,19 @@ biomarkertmle <- function(se,
   type <- match.arg(type)
 
   # ============================================================================
-  # invoke S3 class constructor for "biotmle" object
+  # invoke S4 class constructor for "biotmle" object
   # ============================================================================
-  biotmle <- biotmle(call = call,
-                     topTable = NULL,
-                     limmaOut = NULL,
-                     tmleOut = NULL)
+  biotmle <- .biotmle(
+       SummarizedExperiment(
+          assay = assay(se),
+          rowData = rowData(se),
+          colData = colData(se)
+       ),
+       call = call,
+       tmleOut = NULL,
+       modtestOut = NULL,
+       topTable = NULL
+   )
 
   #=============================================================================
   # set up parallelization based on input
@@ -136,7 +145,7 @@ biomarkertmle <- function(se,
                                     family = family
                                    )
     }
-    biotmle$tmleOut <- as.data.frame(t(biomarkerTMLEout))
+    biotmle@tmleOut <- as.data.frame(t(biomarkerTMLEout))
     return(biotmle)
 
   } else if (type == "outcome") {
@@ -171,7 +180,7 @@ biomarkertmle <- function(se,
                                    family = family
                                   )
     }
-    biotmle$tmleOut <- as.data.frame(biomarkerTMLEout)
+    biotmle@tmleOut <- as.data.frame(biomarkerTMLEout)
     return(biotmle)
 
   } else {
