+++ man/bioTMLE-class.Rd
@@ -1,0 +1,13 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/biotmle.R
+\docType{class}
+\name{bioTMLE-class}
+\alias{bioTMLE-class}
+\alias{.biotmle}
+\title{Constructor for class biotmle}
+\value{
+class \code{biotmle} object, sub-classed from SummarizedExperiment.
+}
+\description{
+Constructor for class biotmle
+}
