+++ man/biomarkertmle.Rd
@@ -35,10 +35,12 @@ fitting the "g" step of the standard TMLE procedure.}
 fitting the "Q" step of the standard TMLE procedure.}
 }
 \value{
-S3 object of class \code{biotmle}, with the \code{tmleOut} and the
-        \code{call} slots containing TMLE-based estimates of the relationship
-        between a biomarker and exposure/outcome variable and the original
-        call to this function (for user reference), respectively.
+S4 object of class \code{biotmle}, generated by sub-classing
+        \code{SummarizedExperiment}, with additional slots containing
+        \code{tmleOut} and \code{call}, among others, containing TMLE-based
+        estimates of the relationship between a biomarker and exposure or
+        outcome variable and the original call to this function (for user
+        reference), respectively.
 }
 \description{
 Computes the causal target parameter defined as the difference between the

[…]
