### 2017-04-04 337cc21
all BioC changes functional
