+++ R/biotmle.R
@@ -1,3 +1,4 @@
+utils::globalVariables(c("new"))
 #' Constructor for class biotmle
 #'
 #' @return class \code{biotmle} object, sub-classed from SummarizedExperiment.
@@ -5,6 +6,7 @@
 #' @importClassesFrom SummarizedExperiment SummarizedExperiment
 #'
 #' @export .biotmle
+#' @exportClass bioTMLE
 #'
 .biotmle <- setClass(
        Class = "bioTMLE",
