+++ R/biomarkertmle.R
@@ -27,7 +27,7 @@ utils::globalVariables("gene")
 #' @importFrom parallel detectCores
 #' @importFrom doParallel registerDoParallel
 #' @importFrom foreach foreach "%dopar%"
-#' @importFrom SummarizedExperiment assay colData
+#' @importFrom SummarizedExperiment assay colData rowData SummarizedExperiment
 #'
 #' @return S4 object of class \code{biotmle}, generated by sub-classing
 #'         \code{SummarizedExperiment}, with additional slots containing
@@ -52,7 +52,7 @@ utils::globalVariables("gene")
 #'
 #' varInt_index <- which(names(colData(illuminaData)) %in% "benzene")
 #'
-#' biomarkerTMLEout <- biomarkertmle(se = illuminaData[1:10, ],
+#' biomarkerTMLEout <- biomarkertmle(se = illuminaData[1:5, ],
 #'                                   varInt = varInt_index,
 #'                                   type = "exposure",
 #'                                   parallel = 1,
@@ -87,14 +87,14 @@ biomarkertmle <- function(se,
   # ============================================================================
   biotmle <- .biotmle(
        SummarizedExperiment(
-          assay = assay(se),
+          assays = assay(se),
           rowData = rowData(se),
           colData = colData(se)
        ),
        call = call,
-       tmleOut = NULL,
-       modtestOut = NULL,
-       topTable = NULL
+       tmleOut = as.data.frame(matrix(NA, 10, 10)),
+       modtestOut = as.data.frame(matrix(NA, 10, 10)),
+       topTable = as.data.frame(matrix(NA, 10, 10))
    )
 
   #=============================================================================
