+++ man/plot_mt.Rd
@@ -1,16 +1,14 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/plots.R
-\name{plot.biotmle}
-\alias{plot.biotmle}
-\title{Plot utility for class biotmle}
+\name{plot_mt}
+\alias{plot_mt}
+\title{Plot p-values from moderated statistical tests for class biotmle}
 \usage{
-\method{plot}{biotmle}(x, ..., type = "pvals_adj")
+plot_mt(biotmle, type = "pvals_adj")
 }
 \arguments{
-\item{x}{object of class \code{biotmle} as produced by an appropriate call to
-\code{biomarkertmle}}
-
-\item{...}{additional arguments passed \code{plot} as necessary}
+\item{biotmle}{object of class \code{biotmle} as produced by an appropriate
+call to \code{biomarkertmle}}
 
 \item{type}{character describing whether to provide a plot of unadjusted or
 adjusted p-values (adjustment performed via Benjamini-Hochberg)}
@@ -39,9 +37,8 @@ varInt_index <- which(names(colData(illuminaData)) \%in\% "benzene")
 designVar <- as.data.frame(colData(illuminaData))[, varInt_index]
 design <- as.numeric(designVar == max(designVar))
 
-limmaTMLEout <- modtest_ic(biotmle = biomarkerTMLEout, IDs = NULL,
-                           design = design)
+limmaTMLEout <- modtest_ic(biotmle = biomarkerTMLEout, design = design)
 
-plot(x = limmaTMLEout, type = "pvals_adj")
+plot_mt(biotmle = limmaTMLEout, type = "pvals_adj")
 
 }
