+++ NAMESPACE
@@ -1,14 +1,17 @@
 # Generated by roxygen2: do not edit by hand
 
-S3method(plot,biotmle)
 export(.biotmle)
 export(biomarkertmle)
 export(heatmap_ic)
 export(modtest_ic)
+export(plot_mt)
 export(volcano_ic)
+exportClasses(bioTMLE)
 importClassesFrom(SummarizedExperiment,SummarizedExperiment)
+importFrom(SummarizedExperiment,SummarizedExperiment)
 importFrom(SummarizedExperiment,assay)
 importFrom(SummarizedExperiment,colData)
+importFrom(SummarizedExperiment,rowData)
 importFrom(doParallel,registerDoParallel)
 importFrom(dplyr,arrange)
 importFrom(dplyr,filter)

[…]
