### 2017-04-04 nhejazi/biotmle#15 @nhejazi
all changes completed as of 337cc21
