### 2017-04-04 02f34f2
Bioc version bump after adding data package
