+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: biotmle
 Title: Targeted Learning for Biomarker Discovery with the Moderated Statistics
-Version: 0.99.4
+Version: 0.99.5
 Authors@R: c(
     person("Nima", "Hejazi", email = "nhejazi@berkeley.edu",
            role = c("aut", "cre", "cph")),
