### 2017-04-04 91209bc
fix vignette and back to plot method
