+++ R/plots.R
@@ -2,10 +2,11 @@
 #'
 #' Histogram of raw or FDR-adjusted p-values from the moderated t-test.
 #'
-#' @param biotmle object of class \code{biotmle} as produced by an appropriate
-#'        call to \code{biomarkertmle}
+#' @param x object of class \code{biotmle} as produced by an appropriate call to
+#'        \code{biomarkertmle}
 #' @param type character describing whether to provide a plot of unadjusted or
 #'        adjusted p-values (adjustment performed via Benjamini-Hochberg)
+#' @param ... additional arguments passed \code{plot} as necessary
 #'
 #' @importFrom ggplot2 ggplot aes geom_histogram geom_point scale_fill_gradientn
 #'             scale_colour_manual guides guide_legend xlab ylab ggtitle
@@ -14,7 +15,7 @@
 #' @return object of class \code{ggplot} containing a histogram of the raw or
 #'         Benjamini-Hochberg corrected p-values (depending on user input).
 #'
-#' @export plot_mt
+#' @export
 #'
 #' @examples
 #' library(dplyr)
@@ -35,17 +36,15 @@
 #'
 #' limmaTMLEout <- modtest_ic(biotmle = biomarkerTMLEout, design = design)
 #'
-#' plot_mt(biotmle = limmaTMLEout, type = "pvals_adj")
+#' plot(x = limmaTMLEout, type = "pvals_adj")
 #'
-plot_mt <- function(biotmle, type = "pvals_adj") {
-
-  stopifnot(class(biotmle) == "bioTMLE")
+plot.bioTMLE <- function(x, ..., type = "pvals_adj") {
 
   pal1 <- wesanderson::wes_palette("Rushmore", 100, type = "continuous")
   pal2 <- wesanderson::wes_palette("Darjeeling", type = "continuous")
 
   if(type == "pvals_raw") {
-    p <- ggplot2::ggplot(as.data.frame(biotmle@topTable), ggplot2::aes(P.Value))
+    p <- ggplot2::ggplot(as.data.frame(x@topTable), ggplot2::aes(P.Value))
     p <- p + ggplot2::geom_histogram(ggplot2::aes(y = ..count..,
       fill = ..count..), colour = "white", na.rm = TRUE, binwidth = 0.025)
     p <- p + ggplot2::ggtitle("Histogram of raw p-values")
@@ -55,7 +54,7 @@ plot_mt <- function(biotmle, type = "pvals_adj") {
   }
 
   if (type == "pvals_adj") {
-    p <- ggplot2::ggplot(as.data.frame(biotmle@topTable),
+    p <- ggplot2::ggplot(as.data.frame(x@topTable),
                          ggplot2::aes(adj.P.Val))
     p <- p + ggplot2::geom_histogram(ggplot2::aes(y = ..count..,
       fill = ..count..), colour = "white", na.rm = TRUE, binwidth = 0.025)
