+++ man/plot.bioTMLE.Rd
@@ -1,14 +1,16 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/plots.R
-\name{plot_mt}
-\alias{plot_mt}
+\name{plot.bioTMLE}
+\alias{plot.bioTMLE}
 \title{Plot p-values from moderated statistical tests for class biotmle}
 \usage{
-plot_mt(biotmle, type = "pvals_adj")
+\method{plot}{bioTMLE}(x, ..., type = "pvals_adj")
 }
 \arguments{
-\item{biotmle}{object of class \code{biotmle} as produced by an appropriate
-call to \code{biomarkertmle}}
+\item{x}{object of class \code{biotmle} as produced by an appropriate call to
+\code{biomarkertmle}}
+
+\item{...}{additional arguments passed \code{plot} as necessary}
 
 \item{type}{character describing whether to provide a plot of unadjusted or
 adjusted p-values (adjustment performed via Benjamini-Hochberg)}
@@ -39,6 +41,6 @@ design <- as.numeric(designVar == max(designVar))
 
 limmaTMLEout <- modtest_ic(biotmle = biomarkerTMLEout, design = design)
 
-plot_mt(biotmle = limmaTMLEout, type = "pvals_adj")
+plot(x = limmaTMLEout, type = "pvals_adj")
 
 }
