+++ NAMESPACE
@@ -1,10 +1,10 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(plot,bioTMLE)
 export(.biotmle)
 export(biomarkertmle)
 export(heatmap_ic)
 export(modtest_ic)
-export(plot_mt)
 export(volcano_ic)
 exportClasses(bioTMLE)
 importClassesFrom(SummarizedExperiment,SummarizedExperiment)
