### 2017-04-04 e0dbc1d
version bump for BioC
