### 2017-04-05 1dceb8d
add examples, fix Bioc reqs
