+++ R/biomarkertmle.R
@@ -52,13 +52,13 @@ utils::globalVariables("gene")
 #'
 #' varInt_index <- which(names(colData(illuminaData)) %in% "benzene")
 #'
-#' biomarkerTMLEout <- biomarkertmle(se = illuminaData[1:5, ],
+#' biomarkerTMLEout <- biomarkertmle(se = illuminaData[1, ],
 #'                                   varInt = varInt_index,
 #'                                   type = "exposure",
 #'                                   parallel = 1,
 #'                                   family = "gaussian",
-#'                                   g_lib = c("SL.mean", "SL.glm"),
-#'                                   Q_lib = c("SL.mean", "SL.glm")
+#'                                   g_lib = c("SL.mean"),
+#'                                   Q_lib = c("SL.mean")
 #'                                  )
 #'
 biomarkertmle <- function(se,
