+++ man/heatmap_ic.Rd
@@ -46,10 +46,9 @@ varInt_index <- which(names(colData(illuminaData)) \%in\% "benzene")
 designVar <- as.data.frame(colData(illuminaData))[, varInt_index]
 design <- as.numeric(designVar == max(designVar))
 
-limmaTMLEout <- modtest_ic(biotmle = biomarkerTMLEout, IDs = NULL,
-                           design = design)
+limmaTMLEout <- modtest_ic(biotmle = biomarkerTMLEout, design = design)
 
 heatmap_ic(x = limmaTMLEout, design = design, FDRcutoff = 0.05,
-           top = 25)
+           top = 15)
 
 }
