+++ man/biomarkertmle.Rd
@@ -61,13 +61,13 @@ colData(illuminaData) <- colData(illuminaData) \%>\%
 
 varInt_index <- which(names(colData(illuminaData)) \%in\% "benzene")
 
-biomarkerTMLEout <- biomarkertmle(se = illuminaData[1:5, ],
+biomarkerTMLEout <- biomarkertmle(se = illuminaData[1, ],
                                   varInt = varInt_index,
                                   type = "exposure",
                                   parallel = 1,
                                   family = "gaussian",
-                                  g_lib = c("SL.mean", "SL.glm"),
-                                  Q_lib = c("SL.mean", "SL.glm")
+                                  g_lib = c("SL.mean"),
+                                  Q_lib = c("SL.mean")
                                  )
 
 }
