+++ DESCRIPTION
@@ -36,6 +36,7 @@ Imports:
     superheat,
     SuperLearner,
     Matrix,
+    stats,
     biotmleData
 Suggests:
     testthat,
