### 2017-04-05 1741db4
bump pkg version and R version for Bioc
