### 2017-04-05 879da7b
fix last Bioc warning + version bump
