+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: biotmle
 Title: Targeted Learning for Biomarker Discovery with the Moderated Statistics
-Version: 0.99.7
+Version: 0.99.8
 Authors@R: c(
     person("Nima", "Hejazi", email = "nhejazi@berkeley.edu",
            role = c("aut", "cre", "cph")),
@@ -31,13 +31,13 @@ Imports:
     ggplot2,
     wesanderson,
     magrittr,
+    stats,
     dplyr,
     SummarizedExperiment,
     superheat,
     SuperLearner,
-    Matrix,
-    stats,
-    biotmleData
+    biotmleData,
+    Matrix
 Suggests:
     testthat,
     rmarkdown,
