### 2017-04-05 ae49187
add rec BiocViews, try build again
