+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: biotmle
 Title: Targeted Learning for Biomarker Discovery with the Moderated Statistics
-Version: 0.99.8
+Version: 0.99.85
 Authors@R: c(
     person("Nima", "Hejazi", email = "nhejazi@berkeley.edu",
            role = c("aut", "cre", "cph")),
@@ -49,5 +49,6 @@ RoxygenNote: 6.0.1
 biocViews:
     GeneExpression,
     DifferentialExpression,
+    Sequencing,
     Microarray,
     RNASeq
