+++ README.md
@@ -6,7 +6,7 @@
 [![Project Status: WIP - Initial development is in progress, but there has not yet been a stable, usable release suitable for the public.](http://www.repostatus.org/badges/latest/wip.svg)](http://www.repostatus.org/#wip)
 [![MIT license](http://img.shields.io/badge/license-MIT-brightgreen.svg)](http://opensource.org/licenses/MIT)
 
-> Targeted learning with the moderated t-statistic for biomarker discovery
+> Targeted Learning with moderated statistics for biomarker discovery
 
 ---
 
@@ -28,18 +28,18 @@ using the approach of moderated statistics implemented in the R package
 ## Installation
 
 - For standard use, install from [Bioconductor](https://bioconductor.org):
-  ```r
+  ```
   source("https://bioconductor.org/biocLite.R")
   biocLite("biotmle")
   ```
 
 - Install the most recent _stable release_ from GitHub:
-  ```r
+  ```
   devtools::install_github("nhejazi/biotmle")
   ```
 
 - To contribute, install the _development version_:
-  ```r
+  ```
   devtools::install_github("nhejazi/biotmle", ref = "develop")
   ```
 
