+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: biotmle
-Title: Targeted Learning for Biomarker Discovery with the Moderated Statistics
+Title: Targeted Learning for Biomarker Discovery with Moderated Statistics
 Version: 0.99.85
 Authors@R: c(
     person("Nima", "Hejazi", email = "nhejazi@berkeley.edu",
