### 2017-04-06 30d1ee6
fix NAMESPACE issue for Bioc
