+++ NAMESPACE
@@ -40,3 +40,6 @@ importFrom(stats,quantile)
 importFrom(superheat,superheat)
 importFrom(tmle,tmle)
 importFrom(wesanderson,wes_palette)
+import(SuperLearner)
+import(biotmleData)
+import(Matrix)
