### 2017-04-06 c3b3a2c
begin adding NGS support
