### 2017-04-07 d06ba7d
Adding RaggedExperiment, basecallQC, CATALYST, GenomicScores, IMAS, semisup, ImpulseDE2, methylInheritance, biotmle, biotmleData, BLMA, TCGAWorkflow, metavizr
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/biotmle@128350 bc3139a8-67e5-0310-9ffc-ced21a209358
