+++ paper/paper.bib
@@ -1,0 +1,37 @@
+@manual{R,
+  title = {R: A Language and Environment for Statistical Computing},
+  author = {{R Core Team}},
+  organization = {R Foundation for Statistical Computing},
+  address = {Vienna, Austria},
+  year = {2017},
+  url = {https://www.R-project.org/}
+}
+
+@book{vdl2011targeted,
+  title={Targeted Learning: Causal Inference for Observational and Experimental
+    Data},
+  author={{van der Laan}, Mark J and Rose, Sherri},
+  year={2011},
+  publisher={Springer Science \& Business Media}
+}
+
+@incollection{smyth2005limma,
+  title={Limma: Linear Models for Microarray Data},
+  author={Smyth, Gordon K},
+  booktitle={Bioinformatics and computational biology solutions using R and
+    Bioconductor},
+  pages={397--420},
+  year={2005},
+  publisher={Springer}
+}
+
+@article{smyth2004linear,
+  title = {Linear Models and Empirical Bayes Methods for Assessing Differential
+    Expression in Microarray Experiments},
+  author = {Smyth, Gordon K},
+  journal={Statistical Applications in Genetics and Molecular Biology},
+  volume={3},
+  number={1},
+  pages={1--25},
+  year={2004}
+}
