+++ R/zzz.R
@@ -1,0 +1,5 @@
+.onAttach <- function(...) {
+  packageStartupMessage("biotmle: Moderated Statistics and Targeted Learning for Biomarker Discovery")
+  packageStartupMessage("Version: ",
+                        utils::packageDescription("biotmle")$Version)
+}
