+++ R/plots.R
@@ -1,0 +1,222 @@
+#' Plot p-values from moderated statistical tests for class biotmle
+#'
+#' Histogram of raw or FDR-adjusted p-values from the moderated t-test.
+#'
+#' @param x object of class \code{biotmle} as produced by an appropriate call to
+#'        \code{biomarkertmle}
+#' @param type character describing whether to provide a plot of unadjusted or
+#'        adjusted p-values (adjustment performed via Benjamini-Hochberg)
+#' @param ... additional arguments passed \code{plot} as necessary
+#'
+#' @importFrom ggplot2 ggplot aes geom_histogram geom_point scale_fill_gradientn
+#'             scale_colour_manual guides guide_legend xlab ylab ggtitle
+#' @importFrom wesanderson wes_palette
+#'
+#' @return object of class \code{ggplot} containing a histogram of the raw or
+#'         Benjamini-Hochberg corrected p-values (depending on user input).
+#'
+#' @export
+#'
+#' @examples
+#' library(dplyr)
+#' library(biotmleData)
+#' library(SummarizedExperiment)
+#' data(illuminaData)
+#' data(biomarkertmleOut)
+#' "%ni%" = Negate("%in%")
+#'
+#' colData(illuminaData) <- colData(illuminaData) %>%
+#'      data.frame %>%
+#'      dplyr::mutate(age = as.numeric(age > median(age))) %>%
+#'      DataFrame
+#'
+#' varInt_index <- which(names(colData(illuminaData)) %in% "benzene")
+#' designVar <- as.data.frame(colData(illuminaData))[, varInt_index]
+#' design <- as.numeric(designVar == max(designVar))
+#'
+#' limmaTMLEout <- modtest_ic(biotmle = biomarkerTMLEout, design = design)
+#'
+#' plot(x = limmaTMLEout, type = "pvals_adj")
+#'
+plot.bioTMLE <- function(x, ..., type = "pvals_adj") {
+
+  pal1 <- wesanderson::wes_palette("Rushmore", 100, type = "continuous")
+  pal2 <- wesanderson::wes_palette("Darjeeling", type = "continuous")
+
+  if(type == "pvals_raw") {
+    p <- ggplot2::ggplot(as.data.frame(x@topTable), ggplot2::aes(P.Value))
+    p <- p + ggplot2::geom_histogram(ggplot2::aes(y = ..count..,
+      fill = ..count..), colour = "white", na.rm = TRUE, binwidth = 0.025)
+    p <- p + ggplot2::ggtitle("Histogram of raw p-values")
+    p <- p + ggplot2::xlab("magnitude of raw p-values")
+    p <- p + ggplot2::scale_fill_gradientn("Count", colors = pal1)
+    p <- p + ggplot2::guides(fill = ggplot2::guide_legend(title = NULL))
+  }
+
+  if (type == "pvals_adj") {
+    p <- ggplot2::ggplot(as.data.frame(x@topTable),
+                         ggplot2::aes(adj.P.Val))
+    p <- p + ggplot2::geom_histogram(ggplot2::aes(y = ..count..,
+      fill = ..count..), colour = "white", na.rm = TRUE, binwidth = 0.025)
+    p <- p + ggplot2::ggtitle("Histogram of BH-corrected FDR p-values")
+    p <- p + ggplot2::xlab("magnitude of BH-corrected p-values")
+    p <- p + ggplot2::scale_fill_gradientn("Count", colors = pal1)
+    p <- p + ggplot2::guides(fill = ggplot2::guide_legend(title = NULL))
+  }
+  return(p)
+}
+
+#==============================================================================#
+## NEXT FUNCTION ===============================================================
+#==============================================================================#
+
[…]
