### 2017-04-07 #156 @nhejazi
@arfon @mgymrek -- The R package has now been fully revised and accepted by Bioconductor (please see the relevant issue [here](https://github.com/Bioconductor/Contributions/issues/254)). Since Bioconductor's review system is quite similar to that of JOSS, the software and paper should now be fully JOSS-compliant.
