### 2017-04-07 #156 @arfon
> Since Bioconductor's review system is quite similar to that of JOSS, the software and paper should now be fully JOSS-compliant.

@nhejazi - we don't have any formal agreement with Bioconductor about this. @karthik - could you advise?
