### 2017-04-08 7626358
minor rnaseq updates
