### 2017-04-08 #156 @nhejazi
@arfon -- I didn't mean anything about a formal agreement: I mentioned that because my previous comment (from 13 March) indicated that the software was being altered, which I assumed would hold up the review. I point out that the software has gone through Bioconductor's entirely separate vetting process for the eventual reviewer (scanning the issue would give them a sense of exactly what standards the software does and does not conform to). My apologies for any confusion!
