### 2017-04-08 #156 @arfon
>  I point out that the software has gone through Bioconductor's entirely separate vetting process for the eventual reviewer (scanning the issue would give them a sense of exactly what standards the software does and does not conform to). My apologies for any confusion.

👍 yep, no problem. I just wanted to make sure that you understood that being accepted into Bioconductor doesn't have an _formal_ meaning here (although I agree it's a very good sign 😁).

BTW, we _do_ have an agreement with the [ROpenSci](https://github.com/ropensci) folks that packages that have been accepted there can be automatically accepted here.
