### 2017-04-08 7678b44
expand biotmle class, fix bugs in rna-seq method
