+++ R/biotmle.R
@@ -20,6 +20,7 @@ utils::globalVariables(c("new"))
 #'           SummarizedExperiment(
 #'              assays = assay(se),
 #'              rowData = rowData(se),
+#'              rowRanges = rowRanges(se),
 #'              colData = colData(se)
 #'           ),
 #'           call = call,
@@ -32,10 +33,13 @@ utils::globalVariables(c("new"))
 #'
 #' example_class <- example_biotmle_class(se = illuminaData)
 #'
+setClassUnion("data.frame_OR_EList", c("data.frame", "EList"))
+setClassUnion("call_char", c("call", "character"))
+
 .biotmle <- setClass(
        Class = "bioTMLE",
-       slots = list(call = "call",
-                    tmleOut = c("data.frame", "EList"),
+       slots = list(call = "call_char",
+                    tmleOut = "data.frame_OR_EList",
                     modtestOut = "data.frame",
                     topTable = "data.frame"),
        contains = "SummarizedExperiment"
