### 2017-04-08 #156 @karthik
@arfon @nhejazi I'm aware of the bioconductor review system but don't know details of their process. But I like the idea of exploring how similar it is to rOpenSci and fast track those submissions if appropriate.  For now let's proceed with this as a regular submission but evaluate before the next bioconductor one comes through.
