+++ R/biomarkertmle.R
@@ -15,6 +15,10 @@ utils::globalVariables("gene")
 #' @param type (character) - choice of the type of TMLE to perform: "exposure"
 #'        to identify biomarkers related to an exposure (input as A), or
 #'        "outcome" to identify biomarkers related to an outcome (input as Y).
+#' @param ngscounts (logical) - whether the data are counts generated from a
+#'        next-generation sequencing (NGS) experiment (e.g., RNA-seq). The
+#'        default setting assumes continuous expression measures as generated by
+#'        microarray-type platforms.
 #' @param parallel (logical, numeric) - whether to use or the number of cores to
 #'        be used when the TMLE-based estimation procedure is parallelized.
 #' @param family (character) - specification of error family: "binomial" or
@@ -64,7 +68,7 @@ utils::globalVariables("gene")
 biomarkertmle <- function(se,
                           varInt,
                           type = c("exposure", "outcome"),
-                          ngsdata = FALSE,
+                          ngscounts = FALSE,
                           parallel = TRUE,
                           family = "gaussian",
                           g_lib = c("SL.glm", "SL.randomForest", "SL.nnet",
@@ -101,7 +105,7 @@ biomarkertmle <- function(se,
   # ============================================================================
   # invoke the voom transform from LIMMA if next-generation sequencing data)
   # ============================================================================
-  if (ngsdata) {
+  if (ngscounts) {
     voom_out <- rnaseq_ic(biotmle)
     voom_exp <- 2^(voom_out$E)
     assay(se) <- voom_exp
@@ -129,7 +133,7 @@ biomarkertmle <- function(se,
   if (type == "exposure") {
 
     # median normalization
-    if (!ngsdata) {
+    if (!ngscounts) {
       Y <- as.data.frame(t(limma::normalizeBetweenArrays(assay(se),
                                                          method = "scale")))
     } else {
@@ -163,7 +167,7 @@ biomarkertmle <- function(se,
                                    )
     }
 
-    if (ngsdata) {
+    if (ngscounts) {
       voom_out$E <- t(biomarkerTMLEout)
       biotmle@tmleOut <- voom_out
     } else {
@@ -177,7 +181,7 @@ biomarkertmle <- function(se,
   #=============================================================================
 
     # median normalization
-    if (!ngsdata) {
+    if (!ngscounts) {
       A <- as.data.frame(t(limma::normalizeBetweenArrays(assay(se),
                                                          method = "scale")))
     } else {
@@ -211,7 +215,7 @@ biomarkertmle <- function(se,
                                   )
     }
 
-    if (ngsdata) {
+    if (ngscounts) {
       voom_out$E <- t(biomarkerTMLEout)
       biotmle@tmleOut <- voom_out
     } else {
