### 2017-04-10 e7cbf6e
fix rna-seq vignette, expand limma-test
