+++ R/modtest_ic.R
@@ -7,7 +7,11 @@
 #' @param biotmle \code{biotmle} object as generated by \code{biomarkertmle}
 #' @param design a design matrix providing the contrasts to be used in the
 #'        linear model fitting procedure of \code{limma::lmFit}
-#' @param ... additional arguments to be passed to functions from \code{limma}
+#' @param adjust the multiple testing correction to be applied to p-values that
+#'        are generated from the moderated tests. The recommended (and default)
+#'        method is that of Benjamini and Hochberg. See \link[limma]{topTable}
+#'        for a list of appropriate methods.
+#' @param ... additional arguments to be passed to \code{limma::lmFit}
 #'
 #' @importFrom limma lmFit eBayes topTable
 #'
@@ -37,15 +41,16 @@
 #'
 modtest_ic <- function(biotmle,
                        design,
+                       adjust = "BH",
                        ...) {
   stopifnot(class(biotmle) == "bioTMLE")
   biomarkerTMLEout <- as.data.frame(biotmle@tmleOut)
 
-  fit <- limma::lmFit(biomarkerTMLEout, design)
-  fit <- limma::eBayes(fit)
+  fit <- limma::lmFit(object = biomarkerTMLEout, design = design, ...)
+  fit <- limma::eBayes(fit = fit)
 
-  tt <- limma::topTable(fit, coef = 1, adjust.method = "BH", sort.by = "none",
-                        number = Inf)
+  tt <- limma::topTable(fit = fit, coef = 1, adjust.method = "BH",
+                        sort.by = "none", number = Inf)
   tt$IDs <- rownames(biomarkerTMLEout)
 
   biotmle@modtestOut <- as.data.frame(fit)
