### 2017-04-10 0e1c08c
fix heatmap for EList objects
