### 2017-04-10 01debbc
add class union doc + fix various errors
