+++ man/data.frame_OR_EList-class.Rd
@@ -1,0 +1,14 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/biotmle.R
+\docType{class}
+\name{data.frame_OR_EList-class}
+\alias{data.frame_OR_EList-class}
+\title{Constructor for class union data.frame_OR_EList}
+\value{
+fusion of classes \code{data.frame} and \code{EList}, used within
+        \code{.biotmle} by class \code{bioTMLE} to handle uncertainty in the
+        object passed to slot "tmleOut".
+}
+\description{
+Constructor for class union data.frame_OR_EList
+}
