+++ man/modtest_ic.Rd
@@ -4,7 +4,7 @@
 \alias{modtest_ic}
 \title{Moderated Statistical Tests for Influence Curves}
 \usage{
-modtest_ic(biotmle, design, ...)
+modtest_ic(biotmle, design, adjust = "BH", ...)
 }
 \arguments{
 \item{biotmle}{\code{biotmle} object as generated by \code{biomarkertmle}}
@@ -12,7 +12,12 @@ modtest_ic(biotmle, design, ...)
 \item{design}{a design matrix providing the contrasts to be used in the
 linear model fitting procedure of \code{limma::lmFit}}
 
-\item{...}{additional arguments to be passed to functions from \code{limma}}
+\item{adjust}{the multiple testing correction to be applied to p-values that
+are generated from the moderated tests. The recommended (and default)
+method is that of Benjamini and Hochberg. See \link[limma]{topTable}
+for a list of appropriate methods.}
+
+\item{...}{additional arguments to be passed to \code{limma::lmFit}}
 }
 \value{
 \code{biotmle} object containing output from \code{limma::lmFit} and
