+++ man/biomarkertmle.Rd
@@ -4,10 +4,10 @@
 \alias{biomarkertmle}
 \title{Biomarker Evaluation with Targeted Minimum Loss-Based Estimation (TMLE)}
 \usage{
-biomarkertmle(se, varInt, type = c("exposure", "outcome"), parallel = TRUE,
-  family = "gaussian", g_lib = c("SL.glm", "SL.randomForest", "SL.nnet",
-  "SL.polymars", "SL.mean"), Q_lib = c("SL.glm", "SL.randomForest", "SL.nnet",
-  "SL.mean"))
+biomarkertmle(se, varInt, type = c("exposure", "outcome"),
+  ngscounts = FALSE, parallel = TRUE, family = "gaussian",
+  g_lib = c("SL.glm", "SL.randomForest", "SL.nnet", "SL.polymars", "SL.mean"),
+  Q_lib = c("SL.glm", "SL.randomForest", "SL.nnet", "SL.mean"))
 }
 \arguments{
 \item{se}{(SummarizedExperiment) - containing expression or next-generation
@@ -21,6 +21,11 @@ corresponding to the treatment or outcome of interest (in the
 \item{type}{(character) - choice of the type of TMLE to perform: "exposure"
 to identify biomarkers related to an exposure (input as A), or
 "outcome" to identify biomarkers related to an outcome (input as Y).}
+
+\item{ngscounts}{(logical) - whether the data are counts generated from a
+next-generation sequencing (NGS) experiment (e.g., RNA-seq). The
+default setting assumes continuous expression measures as generated by
+microarray-type platforms.}
 
 \item{parallel}{(logical, numeric) - whether to use or the number of cores to
 be used when the TMLE-based estimation procedure is parallelized.}
