+++ NAMESPACE
@@ -5,9 +5,12 @@ export(.biotmle)
 export(biomarkertmle)
 export(heatmap_ic)
 export(modtest_ic)
+export(rnaseq_ic)
 export(volcano_ic)
 exportClasses(bioTMLE)
+exportClasses(data.frame_OR_EList)
 importClassesFrom(SummarizedExperiment,SummarizedExperiment)
+importClassesFrom(limma,EList)
 importFrom(SummarizedExperiment,SummarizedExperiment)
 importFrom(SummarizedExperiment,assay)
 importFrom(SummarizedExperiment,colData)
@@ -34,12 +37,11 @@ importFrom(ggplot2,ylab)
 importFrom(limma,eBayes)
 importFrom(limma,lmFit)
 importFrom(limma,topTable)
+importFrom(limma,voom)
+importFrom(limma,voomWithQualityWeights)
 importFrom(magrittr,"%>%")
 importFrom(parallel,detectCores)
 importFrom(stats,quantile)
 importFrom(superheat,superheat)
 importFrom(tmle,tmle)
 importFrom(wesanderson,wes_palette)
-import(SuperLearner)
-import(biotmleData)
-import(Matrix)
