### 2017-04-10 dc7611e
minor changes
