+++ R/biomarkertmle.R
@@ -1,3 +1,5 @@
+utils::globalVariables(c("gene"))
+
 #' Biomarker Evaluation with Targeted Minimum Loss-Based Estimation (TMLE)
 #'
 #' Computes the causal target parameter defined as the difference between the
@@ -152,10 +154,10 @@ biomarkertmle <- function(se,
     }
 
     # perform multi-level TMLE-based estimation for genes as Y
-    biomarkerTMLEout <- foreach::foreach(g = seq_len(ncol(Y)),
+    biomarkerTMLEout <- foreach::foreach(gene = seq_len(ncol(Y)),
                                          .combine = cbind) %dopar% {
       print(paste("Estimating target parameter for", gene, "of", ncol(Y)))
-      out <- biomarkerTMLE_exposure(Y = Y[, g],
+      out <- biomarkerTMLE_exposure(Y = Y[, gene],
                                     W = W,
                                     A = A,
                                     a = unique(A),
@@ -200,12 +202,12 @@ biomarkertmle <- function(se,
     }
 
     # perform multi-level TMLE-bases estimation for genes as A
-    biomarkerTMLEout <- foreach::foreach(g = seq_len(ncol(A)),
+    biomarkerTMLEout <- foreach::foreach(gene = seq_len(ncol(A)),
                                          .combine = rbind) %dopar% {
       print(paste("Estimating causal effect for", gene, "of", ncol(A)))
       out <- biomarkerTMLE_outcome(Y = Y,
                                    W = W,
-                                   A = A[, g],
+                                   A = A[, gene],
                                    a = unique(A),
                                    g_lib = g_lib,
                                    Q_lib = Q_lib,
