### 2017-04-24 fb22ffe
bump x.y.z versions to even y prior to creation of 3_5 branch
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/biotmle@129126 bc3139a8-67e5-0310-9ffc-ced21a209358
