+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: biotmle
 Title: Targeted Learning for Biomarker Discovery with Moderated
         Statistics
-Version: 0.99.90
+Version: 1.0.0
 Authors@R: c( person("Nima", "Hejazi", email = "nhejazi@berkeley.edu",
         role = c("aut", "cre", "cph")), person("Alan", "Hubbard", email
         = "hubbard@berkeley.edu", role = "aut"), person("Weixin",
