### 2017-06-09 #156 @nhejazi
Would it be possible to inquire whether there might be other JOSS editors that might be less busy than the individual assigned? Nothing at all against @mgymrek, just that we haven't heard anything from them at all on this thread since 10 January.
