### 2017-06-16 3e7dc3b
rm modtestOut arg, fix modtest_ic fun
