+++ R/biomarkerTMLE_outcome.R
@@ -12,6 +12,10 @@
 #'        given biomarker.
 #' @param a (numeric vector) - the levels of A against which comparisons are to
 #'        be made.
+#' @param subj_ids (numeric vector) - subject IDs to be passed directly to
+#         \code{tmle::tmle} when there are repeated measures; measurements on
+#'        the same subject should have the exact same numerical identifier;
+#'        coerced to numeric if not provided in the appropriate form.
 #' @param g_lib (char vector) - library of learning algorithms to be used in
 #'        fitting the "g" step of the standard TMLE procedure.
 #' @param Q_lib (char vector) - library of learning algorithms to be used in
@@ -30,6 +34,7 @@ biomarkerTMLE_outcome <- function(Y,
                                   W,
                                   A,
                                   a = 1,
+                                  subj_ids = NULL,
                                   family = "binomial",
                                   g_lib,
                                   Q_lib) {
@@ -44,6 +49,10 @@ biomarkerTMLE_outcome <- function(Y,
   n_a = length(a)
   IC = NULL
   EY = NULL
+  
+  if(!is.null(subj_ids)) {
+    subj_ids <- as.numeric(subj_ids)
+  }
 
   for(i in 1:n_a) {
     A_star = as.numeric(A == a[i])
@@ -53,6 +62,7 @@ biomarkerTMLE_outcome <- function(Y,
                           g.SL.library = g_lib,
                           Q.SL.library = Q_lib,
                           family = family,
+                          id = subj_ids,
                           verbose = FALSE
                          )
     g_0 = fit_tmle$g$g1W
