### 2017-06-16 d1f6160
palettes now consistent with renaming in wesanderson pkg; adds color bounds to volcano plot
