+++ R/plots.R
@@ -40,8 +40,8 @@
 #'
 plot.bioTMLE <- function(x, ..., type = "pvals_adj") {
 
-  pal1 <- wesanderson::wes_palette("Rushmore", 100, type = "continuous")
-  pal2 <- wesanderson::wes_palette("Darjeeling", type = "continuous")
+  pal1 <- wesanderson::wes_palette("Rushmore1", 100, type = "continuous")
+  pal2 <- wesanderson::wes_palette("Darjeeling1", type = "continuous")
 
   if(type == "pvals_raw") {
     p <- ggplot2::ggplot(as.data.frame(x@topTable), ggplot2::aes(P.Value))
@@ -77,6 +77,13 @@ plot.bioTMLE <- function(x, ..., type = "pvals_adj") {
 #'
 #' @param biotmle object of class \code{biotmle} as produced by an appropriate
 #'        call to \code{biomarkertmle}
+#' @param fc_bound (numeric) - indicates the highest magnitude of the fold
+#'        to be colored along the x-axis of the volcano plot; this limits the
+#'        observations to be considered differentially expressed to those in a #'        user-specified interval.
+#' @param pval_bound (numeric) - indicates the largest corrected p-value to be
+#'        colored along the y-axis of the volcano plot; this limits observations
+#'        considered as differentially expressed to those in a user-specified
+#'        interval.
 #'
 #' @importFrom magrittr "%>%"
 #' @importFrom dplyr arrange mutate select filter
@@ -113,24 +120,25 @@ plot.bioTMLE <- function(x, ..., type = "pvals_adj") {
 #'
 #' volcano_ic(biotmle = limmaTMLEout)
 #'
-volcano_ic <- function(biotmle) {
+volcano_ic <- function(biotmle, fc_bound = 3.0, pval_bound = 0.2) {
 
   stopifnot(class(biotmle) == "bioTMLE")
 
-  pal1 <- wesanderson::wes_palette("Rushmore", 100, type = "continuous")
-  pal2 <- wesanderson::wes_palette("Darjeeling", type = "continuous")
+  pal1 <- wesanderson::wes_palette("Rushmore1", 100, type = "continuous")
+  pal2 <- wesanderson::wes_palette("Darjeeling1", type = "continuous")
 
   tt_volcano <- as.data.frame(biotmle@topTable) %>%
     dplyr::arrange(adj.P.Val) %>%
     dplyr::mutate(
       logFC = I(logFC),
       logPval = -log10(P.Value),
-      color = ifelse((logFC > 3.0) & (adj.P.Val < 0.2), "1",
-                      ifelse((logFC < -3.0) & (adj.P.Val < 0.2), "-1", "0"))
+      color = ifelse((logFC > fc_bound) & (adj.P.Val < pval_bound), "1",
+                      ifelse((logFC < -fc_bound) & (adj.P.Val < pval_bound),
+                             "-1", "0"))
     ) %>%
     dplyr::select(which(colnames(.) %in% c("logFC", "logPval", "color"))) %>%
-    dplyr::filter((logFC > stats::quantile(logFC, probs = 0.1)) &
-                   logFC < stats::quantile(logFC, probs = 0.9))
+    dplyr::filter((logFC > stats::quantile(logFC, probs = 0.05)) &
+                   logFC < stats::quantile(logFC, probs = 0.95))
 
   p <- ggplot2::ggplot(tt_volcano, ggplot2::aes(x = logFC, y = logPval))
   p <- p + ggplot2::geom_point(aes(colour = color))
@@ -145,6 +153,7 @@ volcano_ic <- function(biotmle) {
 #==============================================================================#
 ## NEXT FUNCTION ===============================================================
[…]
