### 2017-06-16 #295 @nhejazi
To the reviewer (@NelleV) -- please make sure to evaluate only the `develop` branch of this project, as it contains several important changes that have not yet been moved to `master` (this is because `master` is currently in line with the version of this package currently available on Bioconductor). Thanks!
