### 2017-06-16 b183ff9
trivial documentation fix
