### 2017-06-19 40550c8
make vignettes consistent with new args
