### 2017-06-19 2b0c5f9
add easy tests + test skeletons
