### 2017-06-23 c3c09b8
update man, add tests, fix new errors
