+++ man/heatmap_ic.Rd
@@ -46,9 +46,8 @@ varInt_index <- which(names(colData(illuminaData)) \%in\% "benzene")
 designVar <- as.data.frame(colData(illuminaData))[, varInt_index]
 design <- as.numeric(designVar == max(designVar))
 
-limmaTMLEout <- modtest_ic(biotmle = biomarkerTMLEout, design = design)
+limmaTMLEout <- modtest_ic(biotmle = biomarkerTMLEout)
 
-heatmap_ic(x = limmaTMLEout, design = design, FDRcutoff = 0.05,
-           top = 15)
+heatmap_ic(x = limmaTMLEout, design = design, FDRcutoff = 0.05, top = 15)
 
 }
