+++ man/bioTMLE-class.Rd
@@ -1,21 +1,14 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/biotmle.R
 \docType{class}
-\name{bioTMLE}
-\alias{bioTMLE}
-\alias{bioTMLE}
+\name{bioTMLE-class}
+\alias{bioTMLE-class}
 \alias{.biotmle}
-\title{Constructor for class union data.frame_OR_EList}
+\title{Constructor for class bioTMLE}
 \value{
-fusion of classes \code{data.frame} and \code{EList}, used within
-        \code{.biotmle} by class \code{bioTMLE} to handle uncertainty in the
-        object passed to slot "tmleOut".
-
 class \code{biotmle} object, sub-classed from SummarizedExperiment.
 }
 \description{
-Constructor for class union data.frame_OR_EList
-
 Constructor for class bioTMLE
 }
 \examples{
