### 2017-06-29 6af433e
add Bioc time badge
