### 2017-06-30 dd3001d
fix pkg build errors
