+++ DESCRIPTION
@@ -34,6 +34,7 @@ Imports:
     stats,
     Matrix,
     methods,
+    DBI,
     dplyr,
     SummarizedExperiment,
     superheat,
