+++ NAMESPACE
@@ -1,14 +1,14 @@
 # Generated by roxygen2: do not edit by hand
 
+S3method(plot,bioTMLE)
+export()
 export(.biotmle)
 export(biomarkertmle)
 export(heatmap_ic)
 export(modtest_ic)
-export(plot.bioTMLE)
 export(rnaseq_ic)
 export(volcano_ic)
 exportClasses(bioTMLE)
-exportClasses(data.frame_OR_EList)
 importClassesFrom(SummarizedExperiment,SummarizedExperiment)
 importClassesFrom(limma,EList)
 importFrom(SummarizedExperiment,SummarizedExperiment)
