### 2017-06-30 a97121b
manuall add github ref, minor tweaks
