### 2017-07-01 185f19c
add contribution guide + test for modtest_ic FUN
