### 2017-07-01 838aaab
fix codecov in travis + add github_pat to appveyor
