### 2017-07-01 59b2a38
change github_pat for appveyor builds
