+++ appveyor.yml
@@ -23,7 +23,7 @@ environment:
     WARNINGS_ARE_ERRORS: 1
     USE_RTOOLS: true
     GITHUB_PAT:
-      secure: cc328d7ed854bdb0132bf875a0de03746f7bf84b
+      secure: 057884a468d1109c3992a9ff7b189e20397c5f1a
 
 build_script:
   - travis-tool.sh install_deps
