### 2017-07-04 b368045
remove normalization for outcomes routine
