+++ vignettes/outcomeBiomarkers.Rmd
@@ -1,15 +1,207 @@
 ---
-title: "Identifying Biomarkers from an Outcome Variable"
+title: "Identifying Biomarkers Impacting an Outcome Variable"
 author: "Nima S. Hejazi & Alan E. Hubbard"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Identifying Biomarkers from an Outcome Variable}
+  %\VignetteIndexEntry{Identifying Biomarkers Impacting an Outcome Variable}
   %\VignetteEngine{knitr::rmarkdown}
   \usepackage[utf8]{inputenc}
 ---
 
+```{r core_reqs, echo=FALSE}
+library(parallel)
+library(doParallel)
+library(foreach)
+```
+
+## Data Structure
+
+```{r sim_data}
+# specify key simulation parameters
+set.seed(6423709)
+n <- 100
+g <- 2500
+
+# generate outcome data
+outcome <- rnorm(n = n, mean = 1000, sd = 50)
+
+# generate expression measures related to outcome and those not
+biomarkers_outcome <- replicate(n = n, rnorm(n = g/2, mean = sqrt(outcome), sd = 25))
+biomarkers_unrelated <- replicate(n = n, rnorm(n = g/2, mean = 100, sd = 20))
+biomarkers <- as.data.frame(rbind(biomarkers_outcome, biomarkers_unrelated))
+
+# design matrix (of phenotype-level data) -- no relation to outcomes/biomarkers
+design <- as.data.frame(replicate(4, rnorm(n = n, mean = 10, sd = 5)))
+```
+
+## Biomarker Identification
+
+First, we load the `biotmle` package and some convenience dependencies:
+
 ```{r setup_data}
 library(dplyr)
 library(biotmle)
+library(SummarizedExperiment)
+```
+
+In order to use Targeted Minimum Loss-Based Estimation to identify those
+biomarkers related to the outcome, based on the Averate Treatment Effect (ATE)
+parameter, we need three separate data structures: (1) _W_, baseline covariates
+that could potentially confound the association of biomarkers with the outcome;
+(2) _A_, a matrix of biomarkers, some associated with the outcome of interest
+and others not; and (3) _Y_, the outcome of interest. All values in _W_ and _A_
+ought to be discretized, in order to avoid practical violations of the
+assumption of positivity. To illustrate, we discretize the phenotype-level data
+below:
+
+```{r discrete_W}
+# discretize the phenotype-level data
+W <- design %>%
+  dplyr::mutate(V1 = as.numeric(V1 > median(V1)),
+                V2 = as.numeric(V2 > quantile(V2)[2]),
+                V3 = as.numeric(V3 < quantile(V3)[4]),
+                V4 = as.numeric(V4 > quantile(V4)[1])) %>%
+  DataFrame
+```
+
+To invoke the biomarker identification function (`biomarkertmle`), we also need
+to discretize the biomarkers (again, this ensures that we do not violate the
+assumption of positivity when assessing our target causal parameters, the ATE).
+
+We'll use the following function to assign each of the expression measures in
+the `data.frame` of biomarkers to the quartile of values, doing this over the
+columns (subjects) of the `biomarkers` `data.frame`.
+
+```{r discrete_by_quantile}
+discrete_by_quantile <- function(x) {
+  if( class(x) != "factor" ) {
+    as.numeric(gtools::quantcut(x))
+  } else {
+    as.numeric(x)
+  }
+}
+```
+
+```{r discrete_A}
+biomarkers_discrete <- as.data.frame(sapply(biomarkers, discrete_by_quantile))
+```
+
+```{r make_SE}
[…]
