### 2017-07-04 154ca6d
add encrypted github_pat for appveyor
