### 2017-07-04 cf39329
add nuget install for github_pat decryption on appveyor
