+++ appveyor.yml
@@ -6,10 +6,9 @@ init:
         $ErrorActionPreference = "Stop"
         Invoke-WebRequest http://raw.github.com/krlmlr/r-appveyor/master/scripts/appveyor-tool.ps1 -OutFile "..\appveyor-tool.ps1"
         Import-Module '..\appveyor-tool.ps1'
+
 install:
   ps: Bootstrap
-  - nuget install secure-file -ExcludeVersion
-  - secure-file\tools\secure-file -decrypt path-to\encrypted-filename.ext.enc -secret %GITHUB_PAT%
 
 cache:
   - C:\RLibrary
