### 2017-07-06 0e2bb03
gut outcomes functionality; add misc pkg goodies
