+++ tests/testthat/test_modtest_ic.R
@@ -19,7 +19,6 @@ varInt_index <- which(names(colData(illuminaData)) %in% "benzene")
 
 biomarkerTMLEout <- biomarkertmle(se = illuminaData[1, ],
                                   varInt = varInt_index,
-                                  type = "exposure",
                                   parallel = 1,
                                   family = "gaussian",
                                   g_lib = c("SL.mean"),
