+++ man/biomarkerTMLE_outcome.Rd
@@ -1,45 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/biomarkerTMLE_outcome.R
-\name{biomarkerTMLE_outcome}
-\alias{biomarkerTMLE_outcome}
-\title{TMLE procedure for Biomarker Identication from Outcome}
-\usage{
-biomarkerTMLE_outcome(Y, W, A, a = 1, subj_ids = NULL,
-  family = "binomial", g_lib, Q_lib)
-}
-\arguments{
-\item{Y}{(numeric vector) - a vector of binarized outcome values, thought
-to be impacted by changes in biomarker expression values.}
-
-\item{W}{(numeric matrix) - a matrix of baseline covariates to be controlled
-for in the estimation procedure.}
-
-\item{A}{(numeric vector) - a discretized vector of expression values from a
-given biomarker.}
-
-\item{a}{(numeric vector) - the levels of A against which comparisons are to
-be made.}
-
-\item{subj_ids}{(numeric vector) - subject IDs to be passed directly to
-the same subject should have the exact same numerical identifier;
-coerced to numeric if not provided in the appropriate form.}
-
-\item{family}{(character) - specification of error family: "binomial" or
-"gaussian"}
-
-\item{g_lib}{(char vector) - library of learning algorithms to be used in
-fitting the "g" step of the standard TMLE procedure.}
-
-\item{Q_lib}{(char vector) - library of learning algorithms to be used in
-fitting the "Q" step of the standard TMLE procedure.}
-}
-\value{
-TMLE-based estimate of the relationship between chagnes in biomarker
-        expression and an outcome variable, computed iteratively and saved in
-        the \code{tmleOut} slot in a \code{biotmle} object.
-}
-\description{
-This function performs influence curve-based estimation of the effect of
-expression changes of a biomarker on an outcome while controlling for a set
-of user-specified baseline covariates.
-}
