### 2017-07-07 78731d9
add initial sketch of outcomes functionality
