+++ R/biomarkertmle.R
@@ -12,6 +12,9 @@ utils::globalVariables(c("gene","assay<-"))
 #' @param varInt (numeric) - indicating the column of the design matrix
 #'        corresponding to the treatment or outcome of interest (in the
 #'        "colData" slot of the "se" argument above).
+#' @param type (character) - choice of the type of TMLE to perform: "exposure"
+#'        to identify biomarkers related to an exposure (input as A), or
+#'        "outcome" to identify biomarkers related to an outcome (input as Y).
 #' @param ngscounts (logical) - whether the data are counts generated from a
 #'        next-generation sequencing (NGS) experiment (e.g., RNA-seq). The
 #'        default setting assumes continuous expression measures as generated by
@@ -59,6 +62,7 @@ utils::globalVariables(c("gene","assay<-"))
 #'
 #' biomarkerTMLEout <- biomarkertmle(se = illuminaData[1, ],
 #'                                   varInt = varInt_index,
+#'                                   type = "exposure",
 #'                                   parallel = 1,
 #'                                   family = "gaussian",
 #'                                   g_lib = c("SL.mean"),
@@ -67,6 +71,7 @@ utils::globalVariables(c("gene","assay<-"))
 #'
 biomarkertmle <- function(se,
                           varInt,
+                          type = c("exposure", "outcome"),
                           ngscounts = FALSE,
                           parallel = TRUE,
                           family = "gaussian",
@@ -81,6 +86,11 @@ biomarkertmle <- function(se,
   # catch input and return in output object for user convenience
   # ============================================================================
   call <- match.call(expand.dots = TRUE)
+
+  # ============================================================================
+  # explicitly check that the "type" argument is of the appropriate type
+  # ============================================================================
+  type <- match.arg(type)
 
   # ============================================================================
   # invoke S4 class constructor for "bioTMLE" object
@@ -124,48 +134,97 @@ biomarkertmle <- function(se,
   #=============================================================================
   # TMLE procedure to identify biomarkers based on an EXPOSURE
   # ============================================================================
+  if (type == "exposure") {
 
-  # median normalization
-  if (!ngscounts) {
-    Y <- as.data.frame(t(limma::normalizeBetweenArrays(assay(se),
-                                                       method = "scale")))
+    # median normalization
+    if (!ngscounts) {
+      Y <- as.data.frame(t(limma::normalizeBetweenArrays(assay(se),
+                                                         method = "scale")))
+    } else {
+      Y <- as.data.frame(t(assay(se)))
+    }
+    # simple sanity check of whether Y includes array values
+    if(unique(lapply(Y, class)) != "numeric") {
+      stop("Warning - values in Y do not appear to be numeric...")
+    }
+
+    # exposure / treatment
+    A <- as.numeric(colData(se)[, varInt])
+
+    # baseline covariates
+    W <- as.data.frame(colData(se)[, -varInt])
+    if(dim(W)[2] == 0) {
+      W <- as.numeric(rep(1, length(A)))
+    }
+
+    # perform multi-level TMLE-based estimation for genes as Y
+    biomarkerTMLEout <- foreach::foreach(gene = seq_len(ncol(Y)),
+                                         .combine = cbind) %dopar% {
+      print(paste("Estimating target parameter for", gene, "of", ncol(Y)))
+      out <- biomarkerTMLE_exposure(Y = Y[, gene],
+                                    W = W,
+                                    A = A,
+                                    a = unique(A),
+                                    g_lib = g_lib,
[…]

[…]
