### 2017-07-07 67cf848
add paper Rmd for JOSS
