### 2017-07-07 #295 @nhejazi
Best to consult the `master` branch now. `develop` now contains some functionality that is statistically questionable and still needs to be thought about.
