### 2017-07-13 #295 @NelleV
Hi @nhejazi

I've done a first pass at reviewing the paper (sorry it took so long), and I have a few comments:

- The version number specified in the ticket is wrong, though I don't see where it is specified. The bioconductor version and the github one do match.
- The contribution guidelines are missing. It'd be nice to have a small section in the readme to help new contributors.

There are still a couple of things I need to look at, but I'll get back to you soon.
