### 2017-07-13 9e7442e
minor updates for JOSS
