+++ README.Rmd
@@ -74,6 +74,18 @@ issue](https://github.com/nhejazi/biotmle/issues).
 
 ---
 
+## Contributions
+
+It is our hope that `biotmle` will grow to be adopted as a tool for the
+nonparametric assessment of variable importance in a variety of settings
+involving data generated by experiments in computational biology. To that end,
+contributions are very welcome, though we ask that interested contributors
+consult our [`contribution
+guidelines`](https://github.com/nhejazi/biotmle/blob/master/CONTRIBUTING.md)
+prior to submitting a pull request.
+
+---
+
 ## Citation
 
 After using the `biotmle` R package, please cite it:
