### 2017-07-13 1d40f60
minor tweaks to wording: readme, vignette, docs
