+++ README.Rmd
@@ -76,7 +76,7 @@ issue](https://github.com/nhejazi/biotmle/issues).
 
 ## Contributions
 
-It is our hope that `biotmle` will grow to be adopted as a tool for the
+It is our hope that `biotmle` will grow to be widely adopted as a tool for the
 nonparametric assessment of variable importance in a variety of settings
 involving data generated by experiments in computational biology. To that end,
 contributions are very welcome, though we ask that interested contributors
