+++ docs/articles/rnaseqProcessing.Rmd
@@ -1,10 +1,10 @@
 ---
-title: "Processing and Analyzing RNA-Seq Data with `biotmle`"
+title: "Processing and Analyzing RNA-Seq Data"
 author: "Nima Hejazi & Alan Hubbard"
 date: "`r Sys.Date()`"
 output: rmarkdown::html_vignette
 vignette: >
-  %\VignetteIndexEntry{Processing and Analyzing RNA-Seq Data with `biotmle`}
+  %\VignetteIndexEntry{Processing and Analyzing RNA-Seq Data}
   %\VignetteEngine{knitr::rmarkdown}
   \usepackage[utf8]{inputenc}
 ---
@@ -18,7 +18,9 @@ library(SummarizedExperiment)
 
 ## Introduction
 
-...
+Here, we briefly work through how to use the `biotmle` package with data
+generated by next-generation sequencing technologies, which, in contrast to
+microarray technologies, produce measurements in the form of discrete counts.
 
 ---
 
@@ -43,8 +45,6 @@ design <- as.data.frame(cbind(exp_var, batch, covar))
 head(ngs_data[, 1:7])
 ```
 
-...
-
 ---
 
 ## Processing: Data Transformation
@@ -54,8 +54,6 @@ se <- SummarizedExperiment(assays = list(counts = DataFrame(ngs_data)),
                            colData = DataFrame(design))
 se
 ```
-
-...
 
 ---
 
@@ -81,14 +79,10 @@ data(rnaseqtmleOut)
 head(rnaseqTMLEout@tmleOut$E[,1:6])
 ```
 
-...
-
 ```{r modtest_ic}
 limmaTMLEout <- modtest_ic(biotmle = rnaseqTMLEout)
 head(limmaTMLEout@topTable)
 ```
-
-...
 
 ---
 
