+++ README.md
@@ -52,7 +52,7 @@ If you encounter any bugs or have any specific feature requests, please [file an
 Contributions
 -------------
 
-It is our hope that `biotmle` will grow to be adopted as a tool for the nonparametric assessment of variable importance in a variety of settings involving data generated by experiments in computational biology. To that end, contributions are very welcome, though we ask that interested contributors consult our [`contribution guidelines`](https://github.com/nhejazi/biotmle/blob/master/CONTRIBUTING.md) prior to submitting a pull request.
+It is our hope that `biotmle` will grow to be widely adopted as a tool for the nonparametric assessment of variable importance in a variety of settings involving data generated by experiments in computational biology. To that end, contributions are very welcome, though we ask that interested contributors consult our [`contribution guidelines`](https://github.com/nhejazi/biotmle/blob/master/CONTRIBUTING.md) prior to submitting a pull request.
 
 ------------------------------------------------------------------------
 
