### 2017-07-13 #295 @nhejazi
Hi @NelleV ---

Thanks much for taking the time to go through our `biotmle` package. I've gone ahead and made the suggested changes as of 1d40f60. In regard to package versioning, I've put the current version (1.1.2) of the GitHub package in the `codemeta.json` file, as the current status of the package reflects changes that will be present in the [development version on Bioconductor](https://bioconductor.org/packages/devel/bioc/html/biotmle.html), which will become the stable release when Bioconductor 3.6 becomes the standard in the Fall.
