### 2017-07-19 #295 @karthik
👋 @NelleV  I know you're just back from SciPy and are catching up, but once you are caught up, can you please let me know if you're ready to sign off? 🙏
