### 2017-07-26 #295 @nhejazi
Thanks very much for taking the time to review this package @NelleV! 🙏
@karthik -- thanks very much for your help throughout the review process!

I've archived v1.1.2 of the package and generated a DOI via Zenodo: DOI: 10.5281/zenodo.834849. Please let me know if there's anything more on my end.
