### 2017-07-26 #295 @arfon
@nhejazi - it looks like there are two paper files (`paper.md` and `paper.Rmd`). The later looks to be more closely aligned with our expected format. Could you confirm that the `paper.Rmd` should be the file we should be looking at here? If so, please delete the `paper.md` file.
