### 2017-07-26 3454c6c
rm paper.md, Makefile
