### 2017-07-26 #295 @nhejazi
Confirming that `paper.Rmd` is the appropriate (reproducible) file to be using here. `paper.md` and (now) unnecessary supporting files have been removed as of the latest commit on `master`.
