### 2017-07-26 #295 @arfon
OK thanks @nhejazi. Can you clarify why this is `.Rmd` and not `.md` - our submission tooling is heavily automated and expects to find a file named `paper.md` not `paper.Rmd` 😁
