### 2017-07-26 #295 @nhejazi
Oh, based on a different JOSS submission that I was involved in, the reviewer asked for an `.Rmd` to ensure reproducibility of results. That submission ended up getting stalled on our end, but I went ahead with applying the suggestion to this submission. I should be able to pretty easily convert it to `.md` if you'd like.
