### 2017-07-26 #295 @arfon
>  I should be able to pretty easily convert it to `.md` if you'd like.

👍 Yes please.
