### 2017-07-26 0b83f94
paper Rmd -> paper md + related
