### 2017-07-26 #295 @nhejazi
Done! `.Rmd` -> `.md`, and everything looks close enough to what I had intended.
